requirements.txt
requests
python-dotenv
python-dotenv
beautifulsoup4
flask
google-generativeai
twilio
gTTS
